✨ Show who voted in calendar detail and make week days selectable
- Extract the mobile poll's per-option "who voted what" list into a shared
  VoteSummary component and reuse it in the calendar day detail, so month and
  week views expose the same participant-level breakdown as the grid
- Week view: clicking a day selects it and shows that day's detail below
- Stack the calendar header controls on small screens so the view tabs no
  longer overflow
