+++ apps/web/src/components/poll/vote-summary.tsx
@@ -1,0 +1,109 @@
+"use client";
+
+import { OptimizedAvatarImage } from "@/components/optimized-avatar-image";
+import {
+  filterParticipantsByVote,
+  useParticipants,
+} from "@/components/participants-provider";
+import { useTranslation } from "@/i18n/client";
+
+import VoteIcon from "./vote-icon";
+
+/**
+ * Lists which participants voted yes / if-need-be / no for an option, as avatars
+ * with a vote badge and name. Shared by the mobile poll options and the calendar
+ * view's day detail so both surface the same "who voted what" breakdown.
+ */
+export const VoteSummary: React.FunctionComponent<{ optionId: string }> = ({
+  optionId,
+}) => {
+  const { t } = useTranslation();
+  const { participants } = useParticipants();
+  const participantsWhoVotedYes = filterParticipantsByVote(
+    participants,
+    optionId,
+    "yes",
+  );
+  const participantsWhoVotedIfNeedBe = filterParticipantsByVote(
+    participants,
+    optionId,
+    "ifNeedBe",
+  );
+  const participantsWhoVotedNo = filterParticipantsByVote(
+    participants,
+    optionId,
+    "no",
+  );
+  const noVotes =
+    participantsWhoVotedYes.length + participantsWhoVotedIfNeedBe.length === 0;
+  return (
+    <div>
+      {noVotes ? (
+        <p className="rounded-lg bg-muted p-2 text-center text-muted-foreground text-sm">
+          {t("noVotes")}
+        </p>
+      ) : (
+        <div className="grid grid-cols-2 gap-2">
+          <div className="col-span-1 space-y-2.5">
+            {participantsWhoVotedYes.map(({ name, image }, i) => (
+              // biome-ignore lint/suspicious/noArrayIndexKey: Fix this later
+              <div key={i} className="flex">
+                <div className="relative mr-2.5 flex size-4 items-center justify-center">
+                  <OptimizedAvatarImage
+                    size="sm"
+                    name={name}
+                    src={image ?? undefined}
+                  />
+                  <VoteIcon
+                    type="yes"
+                    size="sm"
+                    className="absolute bottom-0 left-full -translate-x-1 translate-y-1 rounded-full bg-background"
+                  />
+                </div>
+                <div className="truncate text-sm">{name}</div>
+              </div>
+            ))}
+            {participantsWhoVotedIfNeedBe.map(({ name, image }, i) => (
+              // biome-ignore lint/suspicious/noArrayIndexKey: Fix this later
+              <div key={i} className="flex">
+                <div className="relative mr-2.5 flex size-4 items-center justify-center">
+                  <OptimizedAvatarImage
+                    size="sm"
+                    name={name}
+                    src={image ?? undefined}
+                  />
+                  <VoteIcon
+                    type="ifNeedBe"
+                    size="sm"
+                    className="absolute bottom-0 left-full -translate-x-1 translate-y-1 rounded-full bg-background"
+                  />
+                </div>
+                <div className="truncate text-sm"> {name}</div>
+              </div>
+            ))}
+          </div>
+          <div className="col-span-1 space-y-2.5">
+            {participantsWhoVotedNo.map(({ name, image }, i) => (
+              // biome-ignore lint/suspicious/noArrayIndexKey: Fix this later
+              <div key={i} className="flex">
+                <div className="relative mr-2.5 flex size-4 items-center justify-center">
+                  <OptimizedAvatarImage
+                    size="sm"
+                    name={name}
+                    src={image ?? undefined}
+                  />
+                  <VoteIcon
+                    type="no"
+                    size="sm"
+                    className="absolute bottom-0 left-full -translate-x-1 translate-y-1 rounded-full bg-background"
+                  />
+                </div>
+                <div className="truncate text-sm">{name}</div>
+              </div>
+            ))}
+          </div>
+        </div>
+      )}
+    </div>
+  );
+};
